Move out non compute engines
